Added filtering of biom files
Someone you want to provide a mapping file that has samples that are not present in biom file. Now, the biom file will be filtered to include only the samples present in the mapping file, and the rest of the analysis will run as expected. This means you don't need to pre-filter the biom file anymore